made half of map for dataBase

+++ app/src/main/java/com/example/hataru/presentation/forMap/ClusterView.kt
@@ -3,10 +3,9 @@ import android.content.Context
 import android.view.View
 import android.widget.LinearLayout
 import android.widget.TextView
-import androidx.core.content.ContextCompat.getString
 import androidx.core.view.isVisible
 import com.example.hataru.R
-import com.example.hataru.presentation.forMap.flat
+
 
 class ClusterView(context: Context) : LinearLayout(context) {
 
@@ -23,11 +22,11 @@ class ClusterView(context: Context) : LinearLayout(context) {
     }
 
     @SuppressLint("SetTextI18n")
-    fun setData(size : Int, minCost : Int, maxCost : Int) {
+    fun setData(size : Int, minCost : Double, maxCost : Double) {
         flatsCountText.text = size.toString()
         flatsLayout.isVisible = size != 0
-        flatsMinCost.text = context.getString(R.string.min_cost_format, minCost)
-        flatsMaxCost.text = context.getString(R.string.max_cost_format, maxCost)
+        flatsMinCost.text = context.getString(R.string.min_cost_format, minCost.toInt())
+        flatsMaxCost.text = context.getString(R.string.max_cost_format, maxCost.toInt())
     }
 
 }
